refactor: move the sign problem back to f_l part
To keep consistent with Sorella and be convenient.

HISTORY:
Assaraf et al used "wrong" direction for force F,
thus giving a "wrong" sign for \tilde{\psi}_{min},
which happens to be \frac{\partial}{\partial R} instead of
\frac{\partial}{\partial \lambda}.

To fix the problem, we'd either fix the Q or the formula.
To avoid introducing \lambda, we can instead fix the formula,
and make it consistent with Sorella et al